Restructure MasterSuite
